+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/api/config/domain/RoleConfiguration.java
@@ -41,7 +41,7 @@ public interface RoleConfiguration  extends Configurable {
 	
 	@Override
 	default List<String> getPropertyNames() {
-		List<String> properties = new ArrayList<>(); // we are not using superclass properties - enable and order does not make a sence here
+		List<String> properties = new ArrayList<>(); // we are not using superclass properties - enable and order does not make a sense here
 		properties.add(getPropertyName(PROPERTY_DEFAULT_ROLE));
 		properties.add(getPropertyName(PROPERTY_ADMIN_ROLE));
 		return properties;