+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/api/dto/IdmTreeTypeDto.java
@@ -1,13 +1,14 @@
 package eu.bcvsolutions.idm.core.api.dto;
 
-import eu.bcvsolutions.idm.core.api.domain.Codeable;
-import eu.bcvsolutions.idm.core.api.domain.DefaultFieldLengths;
-import eu.bcvsolutions.idm.core.api.domain.Embedded;
+import java.util.UUID;
+
+import javax.validation.constraints.Size;
+
 import org.hibernate.validator.constraints.NotEmpty;
 import org.springframework.hateoas.core.Relation;
 
-import javax.validation.constraints.Size;
-import java.util.UUID;
+import eu.bcvsolutions.idm.core.api.domain.Codeable;
+import eu.bcvsolutions.idm.core.api.domain.DefaultFieldLengths;
 
 /**
  * Type of tree structure
@@ -24,9 +25,13 @@ public class IdmTreeTypeDto extends AbstractDto implements Codeable  {
     @NotEmpty
     @Size(min = 1, max = DefaultFieldLengths.NAME)
     private String name;
-    private boolean defaultTreeType = false; // true, when this type is defined as default organization structure
-    @Embedded(dtoClass = IdmTreeNodeDto.class, enabled = false)
-    private UUID defaultTreeNode;
+    
+    public IdmTreeTypeDto() {
+	}
+    
+    public IdmTreeTypeDto(UUID id) {
+    	super(id);
+   	}
 
     @Override
     public String getCode() {
@@ -43,21 +48,5 @@ public class IdmTreeTypeDto extends AbstractDto implements Codeable  {
 
     public void setName(String name) {
         this.name = name;
-    }
-
-    public boolean isDefaultTreeType() {
-        return defaultTreeType;
-    }
-
-    public void setDefaultTreeType(boolean defaultTreeType) {
-        this.defaultTreeType = defaultTreeType;
-    }
-
-    public UUID getDefaultTreeNode() {
-        return defaultTreeNode;
-    }
-
-    public void setDefaultTreeNode(UUID defaultTreeNode) {
-        this.defaultTreeNode = defaultTreeNode;
     }
 }