+++ Realization/backend/core/core-impl/src/main/java/eu/bcvsolutions/idm/core/model/repository/IdmTreeTypeRepository.java
@@ -1,10 +1,7 @@
 package eu.bcvsolutions.idm.core.model.repository;
 
-import java.util.UUID;
-
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
@@ -21,8 +18,6 @@ import eu.bcvsolutions.idm.core.model.entity.IdmTreeType;
  */
 public interface IdmTreeTypeRepository extends AbstractEntityRepository<IdmTreeType, IdmTreeTypeFilter> {
 	
-	IdmTreeType findOneByCode(@Param("code") String code);
-	
 	@Override
 	@Query(value = "select e from #{#entityName} e" +
 	        " where"
@@ -34,30 +29,10 @@ public interface IdmTreeTypeRepository extends AbstractEntityRepository<IdmTreeT
 	Page<IdmTreeType> find(IdmTreeTypeFilter filter, Pageable pageable);
 	
 	/**
-	 * Returns default tree type
+	 * Finds treeType by code (unique).
 	 * 
+	 * @param code
 	 * @return
 	 */
-	IdmTreeType findOneByDefaultTreeTypeIsTrue();
-	
-	/**
-	 * Clears default tree type for all tree types instead given updatedEntityId
-	 * 
-	 * @param updatedEntityId
-	 */
-	@Modifying
-	@Query("update #{#entityName} e set e.defaultTreeType = false where (:updatedEntityId is null or e.id != :updatedEntityId)")
-	@Deprecated
-	void clearDefaultTreeType(@Param("updatedEntityId") UUID updatedEntityId);
-	
-	/**
-	 * Clear default tree node, when node is deleted etc.
-	 * 
-	 * @param defaultTreeNode
-	 * @return
-	 */
-	@Modifying
-	@Query("update #{#entityName} e set e.defaultTreeNode = null where e.defaultTreeNode.id = :defaultTreeNodeId")
-	@Deprecated
-	int clearDefaultTreeNode(@Param("defaultTreeNodeId") UUID defaultTreeNodeId);
+	IdmTreeType findOneByCode(@Param("code") String code);
 }